src/Services/ClipSearch/requirements.txt
transformers==4.53.0
numpy<2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
fastapi==0.115.14
uvicorn==0.35.0
elasticsearch==8.11.0
pydantic==2.5.3
httpx==0.25.2
python-dotenv==1.0.0
pillow==11.2.1
beautifulsoup4==4.12.3
